+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Allow to change the path for the `git` binary (#1)
     * Added CLI for synchronizing `bin/plugin git-sync sync` (#2)
+    * More security: Git password will now get encrypted and won't load in admin
 1. [](#improved)
     * Wizard: Improved Bitbucket explanation about stripping out `user@` from the copied HTTPS url (#3)
 1. [](#bugfix)